🚚 move `Hook`, `createHook`, `DISCARDED`, and `SKIPPED` to `@datadog/js-core/assembly`
These hook primitives are not browser-specific and belong in the shared
`js-core` package alongside the recently moved utility functions.
