+++ eslint-local-rules/disallowSideEffects.ts
@@ -45,10 +45,10 @@ const pathsWithSideEffect = new Set([
 
 // Those packages are known to have no side effects when evaluated
 const packagesWithoutSideEffect = new Set([
+  '@datadog/js-core/assembly',
   '@datadog/js-core/time',
   '@datadog/js-core/util',
   '@datadog/js-core/monitor',
-  '@datadog/js-core/util',
   '@datadog/browser-core',
   '@datadog/browser-rum-core',
   '@datadog/browser-rum-react/internal',
