+++ packages/browser-core/src/domain/contexts/globalContext.ts
@@ -1,4 +1,4 @@
-import type { Hook } from '../../tools/abstractHooks'
+import type { Hook } from '@datadog/js-core/assembly'
 import { CustomerDataType } from '../context/contextConstants'
 import { storeContextManager } from '../context/storeContextManager'
 import { createContextManager } from '../context/contextManager'
