+++ packages/browser-core/src/domain/contexts/userContext.ts
@@ -1,8 +1,8 @@
 import type { RelativeTime } from '@datadog/js-core/time'
-import type { Hook } from '../../tools/abstractHooks'
+import type { Hook } from '@datadog/js-core/assembly'
+import { SKIPPED } from '@datadog/js-core/assembly'
 import { CustomerDataType } from '../context/contextConstants'
 import { storeContextManager } from '../context/storeContextManager'
-import { SKIPPED } from '../../tools/abstractHooks'
 import { createContextManager } from '../context/contextManager'
 import type { Configuration } from '../configuration'
 import { isEmptyObject } from '../../tools/utils/objectUtils'
