+++ packages/browser-core/src/domain/telemetry/telemetry.ts
@@ -1,5 +1,7 @@
 import { clocksNow } from '@datadog/js-core/time'
 import { getDebugMode, combine } from '@datadog/js-core/util'
+import type { Hook, RecursivePartial } from '@datadog/js-core/assembly'
+import { DISCARDED } from '@datadog/js-core/assembly'
 import type { Context } from '../../tools/serialisation/context'
 import { NO_ERROR_STACK_PRESENT_MESSAGE, isError } from '../error/error'
 import { toStackTraceString } from '../../tools/stackTrace/handlingStack'
@@ -27,8 +29,6 @@ import {
 } from '../../transport'
 import { createIdentityEncoder } from '../../tools/encoder'
 import { createPageMayExitObservable } from '../../browser/pageMayExitObservable'
-import type { Hook, RecursivePartial } from '../../tools/abstractHooks'
-import { DISCARDED } from '../../tools/abstractHooks'
 import { globalObject, isWorkerEnvironment } from '../../tools/globalObject'
 import { noop } from '../../tools/utils/functionUtils'
 import type { TelemetryEvent } from './telemetryEvent.types'
