+++ packages/browser-core/src/index.ts
@@ -180,7 +180,6 @@ export { ErrorHandling, ErrorSource } from './domain/error/error.types'
 export * from './domain/deflate'
 export * from './domain/connectivity'
 export * from './tools/stackTrace/handlingStack'
-export * from './tools/abstractHooks'
 export * from './domain/tags'
 export { correctedChildSampleRate, isSampled, resetSampleDecisionCache, sampleUsingKnuthFactor } from './domain/sampler'
 export { startTelemetrySessionContext } from './domain/contexts/telemetrySessionContext'
