+++ packages/browser-rum-core/src/domain/action/actionCollection.ts
@@ -1,7 +1,8 @@
 import type { RelativeTime, Duration } from '@datadog/js-core/time'
 import type { Observable } from '@datadog/browser-core'
 import { toServerDuration, addDuration } from '@datadog/js-core/time'
-import { noop, SKIPPED } from '@datadog/browser-core'
+import { noop } from '@datadog/browser-core'
+import { SKIPPED } from '@datadog/js-core/assembly'
 import { discardNegativeDuration } from '../discardNegativeDuration'
 import type { RawRumActionEvent } from '../../rawRumEvent.types'
 import { RumEventType } from '../../rawRumEvent.types'
