+++ packages/browser-rum-core/src/domain/contexts/ciVisibilityContext.ts
@@ -1,4 +1,5 @@
-import { display, getInitCookie, SKIPPED } from '@datadog/browser-core'
+import { display, getInitCookie } from '@datadog/browser-core'
+import { SKIPPED } from '@datadog/js-core/assembly'
 import { createCookieObservable } from '../../browser/cookieObservable'
 import type { AssembleHook, DefaultRumEventAttributes } from '../hooks'
 import { SessionType } from './sessionContext'
