+++ packages/browser-rum-core/src/domain/contexts/connectivityContext.spec.ts
@@ -1,6 +1,6 @@
 import { setNavigatorOnLine, setNavigatorConnection } from '@datadog/browser-core/test'
 import type { RelativeTime } from '@datadog/js-core/time'
-import { createHook } from '@datadog/browser-core'
+import { createHook } from '@datadog/js-core/assembly'
 import type { AssembleHook, AssembleHookParams } from '../hooks'
 import { startConnectivityContext } from './connectivityContext'
