+++ packages/browser-rum-core/src/domain/contexts/sessionContext.spec.ts
@@ -1,6 +1,6 @@
 import type { RelativeTime } from '@datadog/js-core/time'
 import { clocksNow } from '@datadog/js-core/time'
-import { DISCARDED, createHook } from '@datadog/browser-core'
+import { DISCARDED, createHook } from '@datadog/js-core/assembly'
 import type { SessionManagerMock } from '@datadog/browser-core/test'
 import { createSessionManagerMock } from '@datadog/browser-core/test'
 import { mockRumConfiguration, noopRecorderApi } from '../../../test'
