+++ packages/browser-rum-core/src/domain/contexts/sourceCodeContext.spec.ts
@@ -1,5 +1,5 @@
 import type { RelativeTime } from '@datadog/js-core/time'
-import { createHook } from '@datadog/browser-core'
+import { createHook } from '@datadog/js-core/assembly'
 import type { AssembleHook, AssembleHookParams } from '../hooks'
 import { registerCleanupTask } from '../../../../browser-core/test'
 import type { RawRumLongAnimationFrameEvent } from '../../rawRumEvent.types'
