+++ packages/browser-rum-core/src/domain/contexts/syntheticsContext.spec.ts
@@ -1,5 +1,5 @@
 import type { RelativeTime } from '@datadog/js-core/time'
-import { createHook } from '@datadog/browser-core'
+import { createHook } from '@datadog/js-core/assembly'
 import { mockSyntheticsWorkerValues } from '../../../../browser-core/test'
 import type { AssembleHook, AssembleHookParams } from '../hooks'
 import { startSyntheticsContext } from './syntheticsContext'
