+++ packages/browser-rum-core/src/domain/contexts/trackingConsentContext.ts
@@ -1,4 +1,4 @@
-import { DISCARDED, SKIPPED } from '@datadog/browser-core'
+import { DISCARDED, SKIPPED } from '@datadog/js-core/assembly'
 import type { TrackingConsentState } from '@datadog/browser-core'
 import type { AssembleTelemetryHook } from '../hooks'
