+++ packages/browser-rum-core/src/domain/contexts/urlContexts.spec.ts
@@ -2,7 +2,7 @@ import { mockClock, registerCleanupTask } from '@datadog/browser-core/test'
 import type { Clock } from '@datadog/browser-core/test'
 import type { RelativeTime } from '@datadog/js-core/time'
 import { clocksNow, relativeToClocks } from '@datadog/js-core/time'
-import { DISCARDED, createHook } from '@datadog/browser-core'
+import { DISCARDED, createHook } from '@datadog/js-core/assembly'
 import { setupLocationObserver } from '../../../test'
 import { LifeCycle, LifeCycleEventType } from '../lifeCycle'
 import type { ViewCreatedEvent } from '../view/trackViews'
