+++ packages/browser-rum-core/src/domain/view/viewCollection.ts
@@ -1,7 +1,8 @@
 import type { Observable } from '@datadog/browser-core'
 import { toServerDuration } from '@datadog/js-core/time'
 import type { Duration, ServerDuration } from '@datadog/js-core/time'
-import { getTimeZone, DISCARDED, isEmptyObject, mapValues } from '@datadog/browser-core'
+import { getTimeZone, isEmptyObject, mapValues } from '@datadog/browser-core'
+import { DISCARDED } from '@datadog/js-core/assembly'
 import { discardNegativeDuration } from '../discardNegativeDuration'
 import type { RecorderApi } from '../../boot/rumPublicApi'
 import type { RawRumViewEvent, ViewPerformanceData } from '../../rawRumEvent.types'
