+++ packages/js-core/assembly/package.json
@@ -1,0 +1,6 @@
+{
+  "private": true,
+  "main": "../cjs/entries/assembly.js",
+  "module": "../esm/entries/assembly.mjs",
+  "types": "../cjs/entries/assembly.d.ts"
+}
