+++ packages/js-core/package.json
@@ -4,6 +4,11 @@
   "license": "Apache-2.0",
   "sideEffects": false,
   "exports": {
+    "./assembly": {
+      "import": "./esm/entries/assembly.mjs",
+      "require": "./cjs/entries/assembly.js",
+      "types": "./cjs/entries/assembly.d.ts"
+    },
     "./time": {
       "import": "./esm/entries/time.mjs",
       "require": "./cjs/entries/time.js",
@@ -24,6 +29,7 @@
     "cjs",
     "esm",
     "src",
+    "assembly",
     "time",
     "monitor",
     "util",
