+++ packages/js-core/src/assembly/index.ts
@@ -1,0 +1,1 @@
+export * from './hook'
