+++ packages/js-core/src/entries/assembly.ts
@@ -1,0 +1,1 @@
+export * from '../assembly'
